Make outputs optional for single samples

+++ modules/local/process/deseq2_qc.nf
@@ -24,14 +24,14 @@ process DESEQ2_QC {
     path clustering_header_multiqc
     
     output:    
-    path "*.pdf"                , emit: pdf
-    path "*.RData"              , emit: rdata
-    path "*pca.vals.txt"        , emit: pca_txt
-    path "*pca.vals_mqc.tsv"    , emit: pca_multiqc
-    path "*sample.dists.txt"    , emit: dists_txt
-    path "*sample.dists_mqc.tsv", emit: dists_multiqc
-    path "*.log"                , emit: log
-    path "size_factors"         , emit: size_factors
+    path "*.pdf"                , optional:true, emit: pdf
+    path "*.RData"              , optional:true, emit: rdata
+    path "*pca.vals.txt"        , optional:true, emit: pca_txt
+    path "*pca.vals_mqc.tsv"    , optional:true, emit: pca_multiqc
+    path "*sample.dists.txt"    , optional:true, emit: dists_txt
+    path "*sample.dists_mqc.tsv", optional:true, emit: dists_multiqc
+    path "*.log"                , optional:true, emit: log
+    path "size_factors"         , optional:true, emit: size_factors
     path  "*.version.txt"       , emit: version
 
     script:
@@ -45,13 +45,15 @@ process DESEQ2_QC {
         --cores $task.cpus \\
         $options.args
     
-    sed "s/deseq2_pca/${label_lower}_deseq2_pca/g" <$pca_header_multiqc >tmp.txt
-    sed -i -e "s/DESeq2 PCA/${label_upper} DESeq2 PCA/g" tmp.txt
-    cat tmp.txt *.pca.vals.txt > ${label_lower}.pca.vals_mqc.tsv
+    if [ -f "R_sessionInfo.log" ]; then
+        sed "s/deseq2_pca/${label_lower}_deseq2_pca/g" <$pca_header_multiqc >tmp.txt
+        sed -i -e "s/DESeq2 PCA/${label_upper} DESeq2 PCA/g" tmp.txt
+        cat tmp.txt *.pca.vals.txt > ${label_lower}.pca.vals_mqc.tsv
 
-    sed "s/deseq2_clustering/${label_lower}_deseq2_clustering/g" <$clustering_header_multiqc >tmp.txt
-    sed -i -e "s/DESeq2 sample/${label_upper} DESeq2 sample/g" tmp.txt
-    cat tmp.txt *.sample.dists.txt > ${label_lower}.sample.dists_mqc.tsv
+        sed "s/deseq2_clustering/${label_lower}_deseq2_clustering/g" <$clustering_header_multiqc >tmp.txt
+        sed -i -e "s/DESeq2 sample/${label_upper} DESeq2 sample/g" tmp.txt
+        cat tmp.txt *.sample.dists.txt > ${label_lower}.sample.dists_mqc.tsv
+    fi
 
     Rscript -e "library(DESeq2); write(x=as.character(packageVersion('DESeq2')), file='${software}.version.txt')"
     """
